requirements.txt
python-telegram-bot==20.6
pytz
apscheduler==3.6.3
